JAVA_OPTS → JAVA_OPTS_APPEND, bump FROM version
The ENV JAVA_OPTS here is a configuration parameter for the base
image ubi8/openjdk-17. In version :1.14, a bug prevented it from
behaving as intended:  JAVA_OPTS is supposed to override all JVM
tuning options from the run script, but it only overrode some of
them.

Version :1.17 and newer of the ubi8/openjdk-17 container image
changes the behaviour of JAVA_OPTS to properly override all run
script parameters.

This was noticed in
<https://github.com/quarkusio/quarkus/issues/35863>

The alternative variable JAVA_OPTS_APPEND is intended to provide
extra parameters on top of those generated by the run script.

Switch to using JAVA_OPTS_APPEND instead of JAVA_OPTS. Take the
opportunity to update to the latest base image version (:1.17).

Signed-off-by: Jonathan Dowland <jdowland@redhat.com>